SPE-3162: Apply red warning tint to bed for out-of-bounds slices (#507)
* Tweaks in BedMaterials.hpp/.cpp

* Fixed alpha channel in function rescaled_with_preserved_ratio()

* Removed unused obsolete code

* Added bool contained_in_bed member variable to struct ProcessorResult

* SPE-3162: Detection of after-slicing bed error state for fdm printers

* Fixed flatpack build

* SPE-3162: Render in red tint the bed instances for which a after-slicing error state has been detected (currently fdm only)

* SPE-3162: Render in red tint the background of thumbnails for bed instances for which a after-slicing error state has been detected (currently fdm only)

* Fixed rendering of bed instances in thumbnails

* SPE-3162: Front-end framework to use red tint for bed instances of sla printers whose after-slicing error state has been detected

* SPE-3162: Detection of after-slicing bed error state for sla printers

* Fix in PlaterScenePresenter::on_bed_instance_updated()

* Ensure ThumbnailStoreUpdater satisfies all thumbnail generation requests + revert of b7a38f259962589f8cee2554930ba8818776e653

* SPE-3162: Reworked logic to detect sla after-slicing error state

* SPE-3162: PlaterScenePresenter set as IFDMResultCacheChangedListener and ISLAResultCacheChangedListener in place of PlaterRenderModule

* Fix after rebase on master

* SPE-3162: Partial rework of thumbnails rendering (use ad-hoc camera and camera trackball)

* Removed code that became obsolete after commit a630ad7b9bd6b11a8bb33d37ab79f892f0803f55

* SPE-3162: Fixed update of bed error state in plater when switching project

* SPE-3162: Fixed after-slice out of bed detection for sla multibeds

* Tweaks in out of bed detection in shaders

* DEBUG_ASSERT() replaced with ASSERT()

* Parameter renaming

+++ src/libslic3r/src/libslic3r/GCode/Thumbnails.cpp
@@ -64,12 +64,12 @@ std::unique_ptr<CompressedImageBuffer> compress_thumbnail_jpg(const Domain::Imag
 {
     // This might be useless, but the C api takes non const ptrs, so better be safe.
     std::vector<unsigned char> rgba_pixels(thumbnail.pixels);
-    const unsigned int row_size = thumbnail.width() * 4;
+    const size_t row_size = thumbnail.width() * 4;
 
     // Store pointers to scanlines start for later use
     std::vector<unsigned char*> rows_ptrs;
     rows_ptrs.reserve(thumbnail.height());
-    for (unsigned int y = 0; y < thumbnail.height(); ++y) {
+    for (size_t y = 0; y < size_t(thumbnail.height()); ++y) {
         rows_ptrs.emplace_back(&rgba_pixels[y * row_size]);
     }
